# Use RUBY_DEBUG instead of NDEBUG