--- a/include/ruby/internal/attr/const.h
+++ b/include/ruby/internal/attr/const.h
@@ -36,11 +36,11 @@
 # define RBIMPL_ATTR_CONST() /* void */
 #endif
 
-/** Enables #RBIMPL_ATTR_CONST iff. #RUBY_NDEBUG. */
-#if RUBY_NDEBUG
-# define RBIMPL_ATTR_CONST_ON_NDEBUG() RBIMPL_ATTR_CONST()
+/** Enables #RBIMPL_ATTR_CONST iff. !RUBY_DEBUG. */
+#if !RUBY_DEBUG
+# define RBIMPL_ATTR_CONST_UNLESS_DEBUG() RBIMPL_ATTR_CONST()
 #else
-# define RBIMPL_ATTR_CONST_ON_NDEBUG() /* void */
+# define RBIMPL_ATTR_CONST_UNLESS_DEBUG() /* void */
 #endif
 
 #endif /* RBIMPL_ATTR_CONST_H */