--- a/ruby_assert.h
+++ b/ruby_assert.h
@@ -8,14 +8,9 @@
  *             modify this file, provided that  the conditions mentioned in the
  *             file COPYING are met.  Consult the file for details.
  */
-#include "ruby/assert.h"
 
-#if !defined(__STDC_VERSION__) || (__STDC_VERSION__ < 199901L)
-/* C89 compilers are required to support strings of only 509 chars. */
-/* can't use RUBY_ASSERT for such compilers. */
-#include <assert.h>
-#else
+#include "ruby/assert.h"
 #undef assert
-#define assert RUBY_ASSERT
-#endif
+#define assert RUBY_ASSERT_NDEBUG
+
 #endif /* RUBY_TOPLEVEL_ASSERT_H */