… include/ruby/internal/arithmetic/char.h
… include/ruby/internal/arithmetic/long.h
… include/ruby/internal/arithmetic/st_data_t.h
… include/ruby/internal/attr/constexpr.h
… include/ruby/internal/core/robject.h
… include/ruby/internal/core/rregexp.h
… include/ruby/internal/core/rstring.h
… include/ruby/internal/fl_type.h