@shyouhei 2020-05-20
Correct me if I'm wrong. What we want to achieve is:

1. You want speed.  Running fast is more important than running right.  No assertions should ever creep into the generated binary, until you are aware of the needs of sanity.
1. You want to control assertions file-by-file.  I don't do that but perhaps you occasionally write `#define RUBY_DEBUG 1` at the top of `vm.c`.  You want that way.
1. You cannot eliminate `assert` to migrate to `RUBY_ASSERT`, because there are submodules using `assert` (see https://github.com/ruby/ruby/pull/3120#issuecomment-631286374).

Given the criteria, I think it is inevitable to either define or undefine `NDEBUG` according to `RUBY_DEBUG`.  Thoughts?