@ko1 2020-05-20
For submodules, we can consider later, IMO.
In other words, this patch disables newly introduce assertions into header files which have huge impact on performance.

The reasons of performance impact are, frequency (RB_TYPE_P, for example) and duplicated assertions (nested functions checks same pre-conditions, for example).