@ko1 2020-05-20
I think the current assertion (pre-condition/post-condition checks and so on) is not enough (e.g. range check for `RARRAY_AREF`). However, to introduce more checks, the performance will be a barrier if enabled by default. And I believe proposed policy (at least for headers) is good to compromise.