@shyouhei 2020-05-20
I have implemented what I thought to be ideal.  You don't like it.  OK, then what do you think is ideal instead?

> In other words, this patch disables newly introduce assertions into header files which have huge impact on performance.

It sounds too short-sighted.  Incremental approach itself isn't bad, but needs a solid goal.  Otherwise the implementation drifts away.

Also (not important):

> The reasons of performance impact are, frequency (RB_TYPE_P, for example) and duplicated assertions (nested functions checks same pre-conditions, for example).

I don't think so.  As I wrote in [ruby-core:98183], assertions themselves are lightweight. I think increased function calling overhead is the key.  Do you have any analysis?